COM.14 - Sellers who did not earn more than 100K
Analyzing salespeople who did not generate 100,000 in revenue in the last quarter is crucial in auditing to identify potential performance issues, market challenges, or inefficiencies in sales strategies. This analysis allows for detecting which salespeople might be struggling and understanding the underlying causes, such as lack of training, motivation, or issues in their sales territories. With these insights, the company can take corrective actions, such as providing additional support, adjusting targets, or revising incentive policies. Additionally, this evaluation helps ensure that sales resources are allocated effectively and that the team is aligned with the company's growth objectives, promoting a culture of high performance and continuous improvement.