COM.09 - Sales to Individuals above 50 thousand
Analyzing sales above 50,000 reais to individuals is essential in auditing to ensure compliance with tax regulations and prevent potential fraudulent activities or money laundering. High-value sales to individuals can indicate significant risks if not properly monitored and verified. This analysis helps ensure that transactions are legitimate, that customers have the financial capacity for such purchases, and that all tax obligations are correctly met. Additionally, it allows the company to identify unusual behavior patterns, protect itself against financial and legal risks, and maintain integrity and transparency in its sales processes.