COM.02 - Sales for Inactive CNPJ
Analyzing sales to inactive CNPJs is crucial in auditing to avoid financial and operational risks associated with lack of diligence in customer management. This analysis helps identify transactions that may lead to financial losses due to default or the customer's inability to fulfill commitments. Additionally, it assists in reviewing credit policies and assessing the financial health of customers, mitigating credit risks and protecting the company against potential losses. By closely monitoring sales to inactive CNPJs, auditing promotes more efficient accounts receivable management and contributes to the financial sustainability of the company.