CP.03 -  Suppliers open for more than 180 days (Domestic or External Market)
It is crucial to analyze suppliers with payments outstanding for more than 180 days, both in the domestic and international markets. This analysis helps identify potential supplier management issues, such as significant delays in payments that can adversely affect the continuous supply of essential products or services. Additionally, it assists in assessing the financial impact of these delays, such as additional costs due to late penalties or loss of discounts for early payment. By identifying suppliers with significantly overdue payments, we can recommend improvements in supplier management policies, more assertive negotiations of payment terms, and enhancements in financial management processes to ensure the stability and operational efficiency of the company.