ETG.01 -  Analysis of goods in bidding
it is crucial to analyze goods in bidding processes. This analysis allows us to assess compliance with bidding requirements, ensuring that the products offered meet the required standards. Additionally, it helps verify the accuracy of quoted prices and identify potential fraudulent practices, such as collusion among suppliers or false information about product quality. By thoroughly analyzing goods in bidding, we can ensure transparency, fairness, and efficiency in the procurement process, promoting more responsible and sustainable resource management for the organization.