FAT.03 -  Difference between Order x Billing dates
Analyzing the difference between Order and Billing dates is crucial for assessing the efficiency and timeliness of sales and product delivery processes. This analysis helps identify delays in order processing, potential operational bottlenecks, and inefficiencies in the supply chain. Additionally, it allows for the verification of whether agreed-upon deadlines with customers are being met, which directly impacts customer satisfaction and the company's reputation. By monitoring the difference between these dates, the company can implement improvements in internal processes, optimize order and billing management, and ensure a more agile and efficient operation.