ETG.05 - Salesperson making sales to other regions
Analyzing sales representatives who conduct sales in other regions is crucial to assess the effectiveness of the company's sales and distribution strategy. This analysis helps identify which sales representatives are expanding the geographic reach of sales and which regions are being more explored. Additionally, it helps understand the specific needs of each regional market, adapting sales and marketing strategies to maximize growth potential in new areas.