CTB.05 -  Analysis of variations in accounting accounts
The analysis of variations in accounting accounts plays an essential role in understanding the financial health and operational performance of an organization. By examining changes in account balances over time, it is possible to identify trends, anomalies, and areas that require special attention. This analysis not only helps detect errors or fraud but also provides valuable insights for strategic financial management. By understanding the reasons behind the variations, managers can make informed decisions that drive operational efficiency, optimize resources, and strengthen the company's competitive position in the market.