CAD.01 -  Use and Consumption Products registered as Saleable
The analysis of errors in the registration of Usage and Consumption products mistakenly classified as Marketable is crucial to ensure the accuracy of records and compliance with accounting and regulatory standards. Identifying and correcting these errors not only ensures the integrity of accounting data but also prevents distortions in inventory information and financial reports. Moreover, this practice promotes more efficient inventory management and better resource allocation, contributing to the optimization of operational processes and meeting customer expectations.