CR.08 -  Titles open for more than 60 days
The analysis of open invoices aged more than 60 days is crucial for a company's financial health. Identifying and monitoring these invoices allows assessing the efficiency of the billing and credit process, as well as anticipating potential cash flow issues. By conducting this analysis, proactive measures can be taken to reduce the risk of delinquency, improve credit management, and strengthen relationships with customers and suppliers.