COM.11.B -  Grouped billing by customer
The analysis of consolidated sales revenue by customer is essential to understand the impact of each customer on the sales volume and total revenue of the company. By evaluating revenue per customer, it's possible to identify key customers who significantly contribute to the company's financial success. This enables the company to focus efforts on customer retention strategies, relationship development, and service customization, thereby improving customer satisfaction and driving sustainable growth.