COM.11.C -  Product billing
The analysis of sales revenue by products is crucial for effective sales and inventory management in a company. By evaluating the performance of each product in terms of generated revenue, it's possible to identify which items are the most profitable and popular among customers. This analysis not only aids in strategic decision-making, such as pricing adjustments and promotions, but also facilitates inventory planning and resource allocation. Additionally, it allows for identifying opportunities to launch new products or discontinue less profitable items, thereby maximizing return on investment.