CR.06 - Discounts Granted on Write-offs in Accounts Receivable
Analyzing discounts granted in the accounts receivable write-off is crucial for ensuring the financial health of the company and the integrity of the collection processes. This analysis helps identify patterns in discount concessions that may negatively impact revenue, as well as detect inconsistent or unauthorized practices that could indicate fraud or operational errors. Additionally, by monitoring discounts, the company can evaluate the effectiveness of its credit and collection policies, making adjustments to optimize cash flow and reduce the risk of bad debt.