TES.02 -  Analysis of movements in the Fixed Fund
The analysis of petty cash transactions within a company is crucial to ensure the transparency and integrity of financial resources allocated for minor expenses. By auditing petty cash, the auditor verifies that funds are being used in accordance with company policies, identifies potential fraud or misuse, and ensures that reimbursements and entries are properly documented and justified. This analysis strengthens confidence in internal controls, preventing the misuse of resources and maintaining consistency in financial management.