CR.16 -  Write-off of credits not linked to banks
Analyzing the write-off of credits not linked to banks is essential for financial and accounting management. It helps identify irregularities or potential fraud, ensures proper recognition of revenue, and evaluates the efficiency of collection processes. This practice also ensures compliance with accounting standards and provides insights into the financial health of the organization, especially regarding customer receivables or other non-bank-related credits.