CP.10 -  Write-off of bills for beneficiary different from Supplier
The analysis of payables written off to beneficiaries different from the registered supplier is critical for identifying potential fraud, operational errors, or payment process failures. This practice ensures that financial resources are directed to the correct recipient as per established contracts. Moreover, it enhances the security of financial transactions, reduces the risk of losses, and ensures compliance with the company's internal and regulatory policies.